Fixed compiler warnings

+++ src/Screen.cpp
@@ -3,13 +3,13 @@
 Screen::Screen(int height, int width, int y0, int x0, const Life& data)
     : m_win{ newwin(height, width, y0, x0), &delwin }, m_displayArr{ data.getData() }
 {
-    if (m_displayArr.size() != height)
+    if (static_cast<int>(m_displayArr.size()) != height)
     {
         throw std::length_error{ "Height must equal array height dimension" };
     }
     for (const auto& arr : m_displayArr)
     {
-        if (arr.size() != width / 2)
+        if (static_cast<int>(arr.size()) != width / 2)
         {
             throw std::length_error{ "Width must equal array width dimension" };
         }
